refactor: split cli generate command
