+++ crates/rulemorph_cli/src/generate.rs
@@ -1,0 +1,65 @@
+use std::path::PathBuf;
+
+use clap::{Args, ValueEnum};
+use rulemorph::{DtoLanguage, generate_dto};
+
+use super::RulesFormatArg;
+use super::input::load_rule;
+use super::output;
+
+#[derive(Args)]
+pub(super) struct GenerateArgs {
+    #[arg(short = 'r', long)]
+    rules: PathBuf,
+    #[arg(long)]
+    rules_format: Option<RulesFormatArg>,
+    #[arg(short = 'l', long)]
+    lang: DtoLanguageArg,
+    #[arg(short = 'n', long)]
+    name: Option<String>,
+    #[arg(short = 'o', long)]
+    output: Option<PathBuf>,
+}
+
+#[derive(Clone, Copy, Debug, ValueEnum)]
+enum DtoLanguageArg {
+    Rust,
+    #[value(alias = "ts")]
+    TypeScript,
+    Python,
+    Go,
+    Java,
+    Kotlin,
+    Swift,
+}
+
+pub(super) fn run(args: GenerateArgs) -> i32 {
+    let (rule, _) = match load_rule(&args.rules, args.rules_format) {
+        Ok(value) => value,
+        Err(code) => return code,
+    };
+
+    let lang = match args.lang {
+        DtoLanguageArg::Rust => DtoLanguage::Rust,
+        DtoLanguageArg::TypeScript => DtoLanguage::TypeScript,
+        DtoLanguageArg::Python => DtoLanguage::Python,
+        DtoLanguageArg::Go => DtoLanguage::Go,
+        DtoLanguageArg::Java => DtoLanguage::Java,
+        DtoLanguageArg::Kotlin => DtoLanguage::Kotlin,
+        DtoLanguageArg::Swift => DtoLanguage::Swift,
+    };
+
+    let output = match generate_dto(&rule, lang, args.name.as_deref()) {
+        Ok(text) => text,
+        Err(err) => {
+            eprintln!("failed to generate dto: {}", err);
+            return 1;
+        }
+    };
+
+    if output::emit_text_output(&output, args.output.as_ref()).is_err() {
+        return 1;
+    }
+
+    0
+}
